added blog detail page, added image and logo, profile page desin

+++ lib/src/infrastructure/repository/data_base_repository.dart
@@ -13,38 +13,51 @@ class DataBaseRepository implements BaseRepository {
 
   @override
   Future<Response> executeRequest(String requestType, String path, header,
-      {body}) async {
+      {body, isMultipar = false}) async {
     Response response;
     var url = Uri.parse(baseUrl + path);
 
     try {
-      switch (requestType) {
-        case "GET":
-          response = await http.get(url, headers: header);
-          break;
-        case "POST":
-          response = await http.post(
-            url,
-            headers: header,
-            body: body,
-          );
-          break;
-        case "PUT":
-          response = await http.put(
-            url,
-            headers: header,
-            body: body,
-          );
-          break;
-        case "DELETE":
-          response = await http.delete(
-            url,
-            headers: header,
-            body: body,
-          );
-          break;
-        default:
-          throw Exception("You did not enter the request type");
+      if (!isMultipar) {
+        switch (requestType) {
+          case "GET":
+            response = await http.get(url, headers: header);
+            break;
+          case "POST":
+            response = await http.post(
+              url,
+              headers: header,
+              body: body,
+            );
+            break;
+          case "PUT":
+            response = await http.put(
+              url,
+              headers: header,
+              body: body,
+            );
+            break;
+          case "DELETE":
+            response = await http.delete(
+              url,
+              headers: header,
+              body: body,
+            );
+            break;
+          default:
+            throw Exception("You did not enter the request type");
+        }
+      } else {
+        try {
+          MultipartRequest _request = http.MultipartRequest("POST", url);
+          MultipartFile _multipartFile =
+              await http.MultipartFile.fromPath('file', body);
+          _request.files.add(_multipartFile);
+          StreamedResponse streamedResponse = await _request.send();
+          response = await Response.fromStream(streamedResponse);
+        } catch (e) {
+          rethrow;
+        }
       }
 
       if (response.statusCode >= 200 && response.statusCode < 300) {
